# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Meridian Advisory API",
    "description": "Consultation booking, payments, client and consultant portals, commerce and content APIs.\n\nAll responses use a single envelope: `{ success: true, data }` or `{ success: false, error: { code, message } }`.\n\nAuthenticate with `Authorization: Bearer <accessToken>`. Access tokens are short-lived; refresh them at `POST /api/auth/refresh`, which reads an httpOnly cookie.",
    "version": "1.0.0",
    "contact": {
      "name": "Support",
      "email": "support@meridianadvisory.co.ke"
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/payments/initialize": {
      "post": {
        "summary": "Start a Paystack checkout",
        "tags": [
          "Payments"
        ],
        "description": "The amount is computed server-side from the booking or order. The request names what is being paid for, never how much.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/payments/verify": {
      "post": {
        "summary": "Verify a transaction with Paystack and settle it",
        "tags": [
          "Payments"
        ],
        "description": "Calls Paystack directly rather than believing the browser. Safe to call repeatedly: settlement is idempotent.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/payments/": {
      "get": {
        "summary": "List payments visible to the caller",
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/payments/{id}": {
      "get": {
        "summary": "Payment detail",
        "tags": [
          "Payments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/payments/refund": {
      "post": {
        "summary": "Refund a settled payment",
        "tags": [
          "Payments"
        ],
        "description": "Requires the payments.refund permission. Partial refunds are supported.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "development"
    }
  ],
...
}